apps/dash-oss-canvas-text/requirements.txt
dash==1.8.0
dash_canvas==0.1.0
Pillow==7.0.0
pytesseract==0.3.2
tesseract-ocr
libtesseract-dev
gunicorn

absl-py==0.7.0
astor==0.7.1
editdistance==0.5.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
Markdown==3.0.1
numpy==1.16.1
opencv-python==4.0.0.21
protobuf==3.6.1
six==1.12.0
tensorboard==1.12.2
tensorflow==1.15.1
termcolor==1.1.0
Werkzeug==1.0.0
